requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.34
python-telegram-bot==22.1
python-dotenv==1.0.1
itsdangerous==2.2.0
waitress==3.0.0